Implementation of Admin Part
A new role, Admin has been added to the application. So, now we have two user types: The customer user, and the admin. CRUD operations have been created for the admin. Now, he can add a new product in the catalog, can see the current products in the catalog, Update the information about current products and delete the product as per the requirement.